Size_t -> long updates for consistency, fixed double free bug, Bestest gemm implemented.

+++ Benchmarking/CuCuBLAS/include/BLASxWrapped.hpp
@@ -10,7 +10,7 @@
 
 void BLASxFlushGPUBuf(short dev_num, int dev_ids[] );
 /// BLASx wrappers for performance evaluation
-double BLASxDgemmWrap(char TransA,  char TransB, size_t M, size_t N, size_t K, double alpha, double* A, size_t ldA, double* B, size_t ldB, double beta, double* C, size_t ldC, size_t T, double cpu_ratio, short dev_num, int dev_ids[]);
-double BLASxExDgemmWrap(char TransA,  char TransB, size_t M, size_t N, size_t K, double alpha, double* A, size_t ldA, double* B, size_t ldB, double beta, double* C, size_t ldC, size_t T, double cpu_ratio, short dev_num, int dev_ids[]);
+double BLASxDgemmWrap(char TransA,  char TransB, long int M, long int N, long int K, double alpha, double* A, long int ldA, double* B, long int ldB, double beta, double* C, long int ldC, long int T, double cpu_ratio, short dev_num, int dev_ids[]);
+double BLASxExDgemmWrap(char TransA,  char TransB, long int M, long int N, long int K, double alpha, double* A, long int ldA, double* B, long int ldB, double beta, double* C, long int ldC, long int T, double cpu_ratio, short dev_num, int dev_ids[]);
 
 #endif
