+++ Benchmarking/CuCuBLAS/src/BLASXWrappers.cpp
@@ -1,7 +1,7 @@
 ///
 /// \author Anastasiadis Petros (panastas@cslab.ece.ntua.gr)
 ///
-/// \brief BLAS lvl 3 wrappers for benchmarks. 
+/// \brief BLAS lvl 3 wrappers for benchmarks.
 ///
 #include <cassert>
 #include <cublasXt.h>
@@ -16,11 +16,11 @@ void BLASxFlushGPUBuf(short dev_num, int dev_ids[] ){
 	BLASx_LRU_free(dev_num, dev_ids);
 }
 
-double BLASxDgemmWrap(char TransA, char TransB, size_t M, size_t N, size_t K, double alpha, double* A, size_t ldA, double* B, size_t ldB, double beta, double* C, size_t ldC, size_t T, double cpu_ratio, short dev_num, int dev_ids[] ){
-	short lvl = 1; 
+double BLASxDgemmWrap(char TransA, char TransB, long int M, long int N, long int K, double alpha, double* A, long int ldA, double* B, long int ldB, double beta, double* C, long int ldC, long int T, double cpu_ratio, short dev_num, int dev_ids[] ){
+	short lvl = 1;
 	double total_t = csecond();
 #ifdef DEBUG
-	lprintf(lvl-1, "|-----> BLASxDgemmWrap(%c,%c,%zu,%zu,%zu,%lf,A(%d),%zu,B(%d),%zu,%lf,C(%d),%zu)\n", 
+	lprintf(lvl-1, "|-----> BLASxDgemmWrap(%c,%c,%zu,%zu,%zu,%lf,A(%d),%zu,B(%d),%zu,%lf,C(%d),%zu)\n",
 		TransA, TransB, M, N, K, alpha, CoCoGetPtrLoc(A), ldA,
 		CoCoGetPtrLoc(B), ldB, beta, CoCoGetPtrLoc(C), ldC);
 #endif
@@ -33,7 +33,7 @@ double BLASxDgemmWrap(char TransA, char TransB, size_t M, size_t N, size_t K, do
 	BLASx_dgemm(CblasColMajor,cpu_op_A,cpu_op_B,M,N,K,alpha,A,ldA, B,ldB, beta,C, ldC, T, dev_num, dev_ids);
 	CoCoSyncCheckErr();
 #ifdef TEST
-	cpu_timer = csecond() - cpu_timer; 
+	cpu_timer = csecond() - cpu_timer;
 	lprintf(lvl, "BLASx execution time -> t_kernel = %lf ms\n", cpu_timer*1000);
 #endif
 
@@ -43,11 +43,11 @@ double BLASxDgemmWrap(char TransA, char TransB, size_t M, size_t N, size_t K, do
 
 }
 
-double BLASxExDgemmWrap(char TransA, char TransB, size_t M, size_t N, size_t K, double alpha, double* A, size_t ldA, double* B, size_t ldB, double beta, double* C, size_t ldC, size_t T, double cpu_ratio, short dev_num, int dev_ids[] ){
-	short lvl = 1; 
+double BLASxExDgemmWrap(char TransA, char TransB, long int M, long int N, long int K, double alpha, double* A, long int ldA, double* B, long int ldB, double beta, double* C, long int ldC, long int T, double cpu_ratio, short dev_num, int dev_ids[] ){
+	short lvl = 1;
 	double total_t = csecond();
 #ifdef DEBUG
-	lprintf(lvl-1, "|-----> BLASxExDgemmWrap(%c,%c,%zu,%zu,%zu,%lf,A(%d),%zu,B(%d),%zu,%lf,C(%d),%zu)\n", 
+	lprintf(lvl-1, "|-----> BLASxExDgemmWrap(%c,%c,%zu,%zu,%zu,%lf,A(%d),%zu,B(%d),%zu,%lf,C(%d),%zu)\n",
 		TransA, TransB, M, N, K, alpha, CoCoGetPtrLoc(A), ldA,
 		CoCoGetPtrLoc(B), ldB, beta, CoCoGetPtrLoc(C), ldC);
 #endif
@@ -60,7 +60,7 @@ double BLASxExDgemmWrap(char TransA, char TransB, size_t M, size_t N, size_t K, 
 	BLASx_gpubuf_dgemm(CblasColMajor,cpu_op_A,cpu_op_B,M,N,K,alpha,A,ldA, B,ldB, beta,C, ldC, T, dev_num, dev_ids);
 	CoCoSyncCheckErr();
 #ifdef TEST
-	cpu_timer = csecond() - cpu_timer; 
+	cpu_timer = csecond() - cpu_timer;
 	lprintf(lvl, "BLASx execution time -> t_kernel = %lf ms\n", cpu_timer*1000);
 #endif
 
@@ -69,5 +69,3 @@ double BLASxExDgemmWrap(char TransA, char TransB, size_t M, size_t N, size_t K, 
 	return total_t;
 
 }
-
-
